+++ public/CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [2017-08-17]
+### Added
+- Site support for White Cloud Pavillion
+
 ### Fixed
 - Userscript now properly recognizes decimal chapters on FallenAngelsScans.